! Remove auto-mapping
This will probably need to be handled very differently in a future API.
Removing it now takes care of yet another macro. The only remaining
macros are:
- TableQuery.apply
- Shape.mapTo
- shouldNotTypecheck
These cannot be replaced with a runtime implementation and they are too
integral to the lifted embedding to avoid their use in all unit tests.